Separated routes from App.js
