+++ src/Categories.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Categories() {
+  return (
+    <div>
+      <h2>Categories</h2>
+      <p>This is the categories page.</p>
+    </div>
+  );
+}
+
+export default Categories;
