+++ src/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Home() {
+  return (
+    <div>
+      <h2>Home</h2>
+      <p>Welcome to the home page!</p>
+    </div>
+  );
+}
+
+export default Home;
